Rename stdst to skilltrees_common
