+++ include/skilltrees/skilltree.h
@@ -5,7 +5,7 @@
 #ifndef INCLUDE_SKILLTREE_H_
 #define INCLUDE_SKILLTREE_H_
 
-#include "skilltrees/stdst.h"
+#include "skilltrees/skilltrees_common.h"
 #include "skilltrees/skill.h"
 #include "skilltrees/inventory.h"
