+++ src/skilltrees/src/skilltrees_common.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
-#include "skilltrees/stdst.h"
+#include "skilltrees/skilltrees_common.h"
 
 /* See stdst.h */
 void** array_element_add(void** array, unsigned int array_len, void* element) {
