+++ tests/skilltrees/CMakeLists.txt
@@ -5,7 +5,7 @@ add_executable(${TEST_EXE}
         inventory_test.c
         skill_test.c
         skilltree_test.c
-        stdst_test.c
+        skilltrees_common_test.c
         main.c)
 
 target_link_libraries(${TEST_EXE} ${CRITERION_LIBRARY})
